refactor: VendingMachine 인스턴스 생성 전에 출력하기 위해 CoinCount를 통해 Map을 객체로 감쌌다.
