+++ src/main/java/vendingmachine/domain/VendingMachine.java
@@ -1,6 +1,5 @@
 package vendingmachine.domain;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -9,14 +8,14 @@ public class VendingMachine {
     private static final String ERROR_HEADER = "[ERROR] ";
     private static final String NOT_EXIST_PRODUCT = "구매할 수 있는 상품이 존재하지 않습니다. 자판기에서 구매할 수 있고 존재하는 상품을 입력해주세요. ";
 
-    private final Map<Coin, Integer> coinCount = new HashMap<>();
+    private final CoinCount coinCount;
     private final List<Product> products;
     private int inputMoney;
 
-    public VendingMachine(List<Product> products, int inputMoney, int vendingMachineMoney) {
+    public VendingMachine(CoinCount coinCount, List<Product> products, int inputMoney) {
+        this.coinCount = coinCount;
         this.products = products;
         this.inputMoney = inputMoney;
-        createRandomCoinList(vendingMachineMoney);
     }
 
     public void buyProduct(String inputProductName) {
@@ -24,6 +23,10 @@ public class VendingMachine {
         inputMoney -= product.getPrice();
         product.buyOneProduct();
 
+    }
+
+    public CoinCount getCoinCount() {
+        return coinCount;
     }
 
     public int getInputMoney() {
@@ -48,30 +51,6 @@ public class VendingMachine {
         return minValue;
     }
 
-    private void createRandomCoinList(int vendingMachineMoney) {
-        for (Coin coinName : Coin.values()) {
-            coinCount.put(coinName, 0);
-        }
-
-        Integer randomCoin;
-        while (vendingMachineMoney != 0) {
-            randomCoin = Coin.getRandomAmount();
-            if (randomCoin <= vendingMachineMoney) {
-                vendingMachineMoney -= randomCoin;
-                Coin randomCoinName = findCoinByValue(randomCoin);
-                coinCount.replace(randomCoinName, coinCount.get(randomCoinName) + 1);
-            }
-        }
-    }
-
-    private Coin findCoinByValue(int coinValue) {
-        for (Coin coin : Coin.values()) {
-            if (coin.getAmount() == coinValue) {
-                return coin;
-            }
-        }
-        return null;
-    }
 
     private Product findProudctInVendingMachine(String productName) {
         for (Product product : products) {
